+++ doctrine.composer.json
@@ -22,16 +22,16 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^7.5",
-        "petrknap/doctrine-ormmappingnamingstrategy": "*"
+        "petrknap/doctrine-ormnamingstrategy": "*"
     },
     "autoload": {
         "psr-4": {
-            "PetrKnap\\Doctrine\\OrmMappingNamingStrategy\\": "packages/Doctrine/OrmMappingNamingStrategy/src"
+            "PetrKnap\\Doctrine\\OrmNamingStrategy\\": "packages/Doctrine/OrmNamingStrategy/src"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "PetrKnap\\Doctrine\\OrmMappingNamingStrategy\\Test\\": "packages/Doctrine/OrmMappingNamingStrategy/tests"
+            "PetrKnap\\Doctrine\\OrmNamingStrategy\\Test\\": "packages/Doctrine/OrmNamingStrategy/tests"
         }
     }
 }
