+++ packages/Doctrine/OrmNamingStrategy/.gitattributes
