+++ packages/Doctrine/OrmNamingStrategy/LICENSE
