+++ packages/Doctrine/OrmNamingStrategy/composer.json
@@ -6,8 +6,8 @@
         "doctrine/orm": "^2.3"
     },
     "WARNING": "This file is updated automatically. All keys will be overwritten, except of 'description', 'keywords', 'require' and 'conflict'.",
-    "name": "petrknap/doctrine-ormmappingnamingstrategy",
-    "homepage": "https://petrknap.github.io/docs/doctrine-ormmappingnamingstrategy.html",
+    "name": "petrknap/doctrine-ormnamingstrategy",
+    "homepage": "https://petrknap.github.io/docs/doctrine-ormnamingstrategy.html",
     "license": "MIT",
     "authors": [
         {
@@ -20,13 +20,13 @@
     },
     "autoload": {
         "psr-4": {
-            "PetrKnap\\Doctrine\\OrmMappingNamingStrategy\\": "src"
+            "PetrKnap\\Doctrine\\OrmNamingStrategy\\": "src"
         },
         "files": []
     },
     "autoload-dev": {
         "psr-4": {
-            "PetrKnap\\Doctrine\\OrmMappingNamingStrategy\\Test\\": "tests"
+            "PetrKnap\\Doctrine\\OrmNamingStrategy\\Test\\": "tests"
         }
     }
 }
